+++ app-arducon/src/main/java/com/keelim/arducon/ui/screen/main/MainViewModel.kt
@@ -2,8 +2,8 @@ package com.keelim.arducon.ui.screen.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.data.repository.ArduconRepository
 import com.keelim.model.DeepLink
 import com.keelim.scheme.notification.SchemeNotificationManager
