+++ app-arducon/src/main/java/com/keelim/arducon/ui/screen/ogtag/OgTagPreviewViewModel.kt
@@ -2,8 +2,8 @@ package com.keelim.arducon.ui.screen.ogtag
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.launch
