+++ core/common/src/main/java/com/keelim/common/di/DispatchersModule.kt
@@ -1,11 +1,10 @@
-package com.keelim.core.network.di
+package com.keelim.common.di
 
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers.DEFAULT
-import com.keelim.core.network.KeelimDispatchers.IO
-import com.keelim.core.network.KeelimDispatchers.MAIN
-import com.keelim.core.network.KeelimDispatchers.MAIN_IMMEDIATE
-import com.keelim.core.network.KeelimDispatchers.UNCONFINED
+import com.keelim.common.KeelimDispatchers.DEFAULT
+import com.keelim.common.KeelimDispatchers.IO
+import com.keelim.common.KeelimDispatchers.MAIN
+import com.keelim.common.KeelimDispatchers.MAIN_IMMEDIATE
+import com.keelim.common.KeelimDispatchers.UNCONFINED
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -21,30 +20,30 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 object DispatchersModule {
     @Provides
-    @Dispatcher(IO)
+    @com.keelim.common.Dispatcher(IO)
     fun providesIODispatcher(): CoroutineDispatcher = Dispatchers.IO
 
     @Provides
-    @Dispatcher(MAIN)
+    @com.keelim.common.Dispatcher(MAIN)
     fun providesMainDispatcher(): CoroutineDispatcher = Dispatchers.Main
 
     @Provides
-    @Dispatcher(MAIN_IMMEDIATE)
+    @com.keelim.common.Dispatcher(MAIN_IMMEDIATE)
     fun providesMainImmediateDispatcher(): CoroutineDispatcher = Dispatchers.Main.immediate
 
     @Provides
-    @Dispatcher(DEFAULT)
+    @com.keelim.common.Dispatcher(DEFAULT)
     fun providesDefaultDispatcher(): CoroutineDispatcher = Dispatchers.Default
 
     @Provides
-    @Dispatcher(UNCONFINED)
+    @com.keelim.common.Dispatcher(UNCONFINED)
     fun providesUnconfinedDispatcher(): CoroutineDispatcher = Dispatchers.Unconfined
 
     @Provides
     @Singleton
     @ApplicationScope
     fun providesCoroutineScope(
-        @Dispatcher(IO) dispatcher: CoroutineDispatcher,
+        @com.keelim.common.Dispatcher(IO) dispatcher: CoroutineDispatcher,
     ): CoroutineScope = CoroutineScope(SupervisorJob() + dispatcher)
 }
