+++ core/data/src/main/java/com/keelim/core/data/source/HistoryRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.keelim.core.data.source
 
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
+import com.keelim.common.di.ApplicationScope
 import com.keelim.core.database.mapper.toSimpleHistoryModel
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers
-import com.keelim.core.network.di.ApplicationScope
 import com.keelim.data.repository.HistoryRepository
 import com.keelim.shared.data.database.dao.HistoryDao
 import com.keelim.shared.data.database.dao.TimerHistoryDao
