+++ core/data/src/main/java/com/keelim/core/data/source/TimeRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.keelim.core.data.source
 
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.time.TrustedTimeClient
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.data.repository.TimeRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.tasks.await
