+++ core/data/src/main/java/com/keelim/core/data/source/note/NoteRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.keelim.core.data.source.note
 
+import com.keelim.common.Dispatcher
+import com.keelim.common.KeelimDispatchers
 import com.keelim.core.database.mapper.toNotices
 import com.keelim.core.database.mapper.toNoticesEntity
-import com.keelim.core.network.Dispatcher
-import com.keelim.core.network.KeelimDispatchers
 import com.keelim.data.repository.NoteRepository
 import com.keelim.model.Notices
 import com.keelim.shared.data.database.dao.NoteDao
